% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adult}
\alias{adult}
\title{Adult Data Set}
\format{
\describe{
\item{age}{The individual's age.}
\item{workclass}{The individual's type of employment. Can include Private, Federal-gov, Without-pay, etc.}
\item{fnlwgt}{Final weight.}
\item{education}{The individual's education level. Can include Bachelors, Some-college, etc.}
\item{education_num}{A numeric representation of the individual's education level.}
\item{marital_status}{The individual's marital status. Can include Never-married, Divorced, Widowed, etc.}
\item{occupation}{The individual's occupation industry. Can include Tech-support, Farming-fishing, etc.}
\item{relationship}{The individual's relationship to other family members. Can include Wife, Husband, Own-child, etc.}
\item{race}{The individual's race. Can include White, Asian-Pac-Islander, etc.}
\item{sex}{The individual's sex, male or female.}
\item{capital_gain}{The individual's capital gain.}
\item{capital_loss}{The individual's capital loss.}
\item{hours_per_week}{The individual's hours worked per week.}
\item{native_country}{The individual's native country.}
\item{salary}{The individual's salary, either <=50K or >50K.}}
}
\source{
https://archive.ics.uci.edu/ml/machine-learning-databases/adult/
}
\usage{
adult
}
\description{
Adult Data Set
}
\examples{
\dontrun{
adult \%>\%
filter(salary = "<=50K")}
\dontrun{
adult \%>\%
group_by(marital_status) \%>\%
count()}
}
\keyword{datasets}
