% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{args_UCI}
\alias{args_UCI}
\title{Arguments Data Set}
\format{
\describe{
\item{name}{The name of the dataset.}
\item{links}{The link to the dataset provides data arguement for the function. Data argument is located below the index header inside the "Data Folder".}
\item{webpage}{The webpage needed for the webpage argument in read_UCI function. Webpage is ../machine-learning-databases/"webpage"/.}
}
}
\usage{
args_UCI
}
\description{
This data set provides the user with the arguments needed to import data from the UCI repository using the read_UCI function.
}
\keyword{datasets}
