% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{breast_cancer}
\alias{breast_cancer}
\title{Breast Cancer Data Set}
\format{
A dataframe with 286 instances and 10 variables:
\describe{
\item{class}{The reoccurance of events.}
\item{age}{The age of the patient at the time of diagnosis.}
\item{menopause}{Whether the patient is pre- or postmenopausal at time of diagnosis.}
\item{inv_nodes}{The number (range 0 - 39) of axillary lymph nodes that contain metastatic breast cancer visible on histological examination.}
\item{node_caps}{The presence of cancer in the lymph nodes.}
\item{deg_malig}{The Degree of malignancy, the histological grade (range 1-3) of the tumor.}
\item{breast}{The prescence of brest cancer in either breast.}
\item{breast_quad}{Breast quadrant, the breast may be divided into four quadrants, using the nipple as a central point.}
\item{irradiat}{Irradiation, the use of radiation therapy}
}
}
\source{
\url{http://archive.ics.uci.edu/ml/datasets/Breast+Cancer}
}
\usage{
breast_cancer
}
\description{
Real patient data obtained from the Institute of Oncology, Ljubljana
}
\keyword{datasets}
