% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{car}
\alias{car}
\title{Car Evalutaion Data Set}
\format{
A data frame with 1728 rows and 7 variables:
\describe{
\item{buying}{The buying price of a car.}
\item{maint}{The price of maintenence.}
\item{doors}{The number of car doors.}
\item{persons}{The capacity of people to carry.}
\item{lug_boot}{The size of the luggage boot.}
\item{safety}{The estimated safety of the car.}
\item{class_values}{The car class values.}
}
}
\source{
\url{http://archive.ics.uci.edu/ml/datasets/Car+Evaluation}
}
\usage{
car
}
\description{
Car Evalutaion Data Set
}
\keyword{datasets}
