% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flags}
\alias{flags}
\title{Flags Data Set}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 194 rows and 30 columns.}
\source{
http://archive.ics.uci.edu/ml/datasets/Flags
}
\usage{
flags
}
\description{
Flags Data Set
}
\examples{
\dontrun{
flags \%>\%
filter(bars >= 3)}
\dontrun{
flags \%>\%
group_by(mainhue) \%>\%
count()}
}
\keyword{datasets}
