% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{italian}
\alias{italian}
\title{Italian Restaurants dataset}
\format{\describe{
\item{restaurant}{name of restaurant}
\item{price}{price of a meal for two with drinks, U.S. dollars}
\item{food}{food rating, scale 0-30}
\item{decor}{decor rating, scale 0-30}
\item{service}{service rating, scale 0-30}
\item{east}{indicator of whether the restaurant is east of 5th Avenue}}}
\source{
\link{http://gattonweb.uky.edu/sheather/book/docs/datasets/nyc.csv}
}
\usage{
italian
}
\description{
Italian Restaurants dataset
}
\examples{
#Calculate the mean price
mean(italian$price)
}
\keyword{datasets}
