% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{las_vegas}
\alias{las_vegas}
\title{Las Vegas Data Set}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 504 rows and 20 columns.}
\source{
https://archive.ics.uci.edu/ml/datasets/Las+Vegas+Strip
}
\usage{
las_vegas
}
\description{
Las Vegas Data Set
}
\examples{
\dontrun{
las_vegas \%>\%
filter(score == 5)}
\dontrun{
adult \%>\%
group_by(hotel_stars) \%>\%
count()}
}
\keyword{datasets}
