% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{liver}
\alias{liver}
\title{Liver Disorders Data Set}
\format{
A dataframe with 214 instances and 11 variables:
\describe{
\item{mcv}{Mean corpuscular volume.}
\item{alkphos}{Alkphos alkaline phosphotase.}
\item{sgpt}{Sgpt alanine aminotransferase.}
\item{sgot}{Sgot aspartate aminotransferase.}
\item{gammagt}{Gamma-glutamyl transpeptidase.}
\item{drinks}{Number of half-pint equivalents of alcoholic beverages drunk per day.}
\item{selector}{Field created by the BUPA researchers to split the data into train/test sets.}
}
}
\source{
\url{http://archive.ics.uci.edu/ml/datasets/Liver+Disorders}
}
\usage{
liver
}
\description{
Liver Disorders Data Set
}
\keyword{datasets}
