% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_UCI.R
\name{read_UCI}
\alias{read_UCI}
\title{read_UCI}
\usage{
read_UCI(webpage, data_file, names_file = NA, data_delim = ",",
  data_col_names = FALSE, names_delim = ",", names_col_names = FALSE,
  ...)
}
\arguments{
\item{webpage}{A string that corresponds with the the name of the dataset as
it appears in the \href{https://archive.ics.uci.edu/ml/machine-learning-databases/}{Parent Directory}.
Note, you do not need to include "/", and correct spelling is important.}

\item{data_file}{A string that corresponds with the primary file name you
wish to read in (typically data). This is visible on the page following the
parent directory. You can also navigate to the page by selecting "Data
Folder" on the homepage of the dataset. Correct spelling is important. If
there is a file extension such as \code{.csv} or \code{.xlsx}, that needs to be
included.}

\item{names_file}{Default is NA. A string that corresponds with the secondary
file name you wish to read in (typically names). This is visible on the
page following the parent directory. You can also navigate to the page by
selecting "Data Folder" on the homepage of the dataset. Correct spelling is
important. If there is a file type such as \code{.csv}, that needs to be
included.}

\item{data_delim}{Default is ",". A single character delimiter used to
separate fields within the data file.}

\item{data_col_names}{Default is \code{FALSE}. Either \code{TRUE}, \code{FALSE}, or a
character vector of column names for the data file.}

\item{names_delim}{Default is ",". A single character delimiter used to
separate fields within the names file.}

\item{names_col_names}{Default is \code{FALSE}. Either \code{TRUE}, \code{FALSE}, or a
character vector of column names for the names file.}

\item{...}{Arguments to be based to \code{openxlsx::read.xlsx} (if ".xls"
is detected) or to \code{readr::read_delim}.}
}
\description{
read_UCI
}
\examples{
#Read in the Las Vegas Trip Advisor Reviews dataset
las_vegas <- read_UCI("00397", "LasVegasTripAdvisorReviews-Dataset.csv", data_delim = ";")
#Read in Iris dataset
iris_uci <- read_UCI("iris", "iris.data")
}
