% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_UCI.R
\name{search_UCI}
\alias{search_UCI}
\title{Read files from the UCI Machine Learning Repository}
\usage{
search_UCI(
  data_type = NULL,
  data_task = NULL,
  data_area = NULL,
  min_instances = 0
)
}
\arguments{
\item{data_type}{A string that corresponds with the the name of the dataset as
it appears in the \href{https://archive.ics.uci.edu/ml/machine-learning-databases/}{Parent Directory}.
Note, you do not need to include "/", and correct spelling is important.}

\item{data_task}{A string that corresponds with the primary file name you
wish to read in (typically data). This is visible on the page following the
parent directory. You can also navigate to the page by selecting "Data
Folder" on the homepage of the dataset. Correct spelling is important. If
there is a file extension such as \code{.csv} or \code{.xlsx}, that needs to be
included.}

\item{data_area}{Default is ",". A single character delimiter used to
separate fields within the data file.}

\item{min_instances}{Default is \code{FALSE}. Either \code{TRUE}, \code{FALSE}, or a
character vector of column names for the data file.}
}
\description{
Read files from the UCI Machine Learning Repository
}
\examples{
dataset_pref1 <- search_UCI("Time-Series", "Classification", "Life Sciences", 300)
dataset_pref2 <- search_UCI(data_type = "Text", data_area = "Business")

}
