% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_train_split.R
\name{test_train_split}
\alias{test_train_split}
\title{test_train_split}
\usage{
test_train_split(data, test_size, y_var)
}
\arguments{
\item{data}{An existing data frame that has been pulled in from the UCI database.}

\item{test_size}{An integer that describes the size of the test set as the number of rows in the new test set.}

\item{y_var}{The response variable of interest. Argument requires y_var variable as a string}
}
\description{
test_train_split
}
\examples{
# create test and train data sets from iris
iris_datasets <- test_train_split(iris_uci, 20)
iris_datasets$test_set
iris_datasets$train_set
iris_datasets$test_input
iris_datasets$test_output
iris_datasets$train_input
iris_datasets$train_output
}
