% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tictactoe}
\alias{tictactoe}
\title{Tic-Tac-Toe Data Set}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 958 rows and 10 columns.}
\source{
http://archive.ics.uci.edu/ml/datasets/Tic-Tac-Toe+Endgame
}
\usage{
tictactoe
}
\description{
Tic-Tac-Toe Data Set
}
\examples{
\dontrun{
tictactoe \%>\%
filter(class == "positive")}
\dontrun{
tictactoe \%>\%
group_by(center) \%>\%
count()}
}
\keyword{datasets}
