% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wine}
\alias{wine}
\title{Wine Data Set}
\format{
A data frame with 178 rows and 14 variables:
\describe{
\item{class}{Type of wine.}
\item{alcohol}{Alcohol.}
\item{malic_acid}{Malic acid.}
\item{ash}{Ash.}
\item{ash_alcalinity}{Ash alcalinity.}
\item{magnesium}{Magnesium.}
\item{total_phenols}{Total Phenols}
\item{flavanoids}{Flavanoids.}
\item{non_flavanoid_phenols}{Non-flavanoid phenols.}
\item{proanthocyanins}{Proanthocyanins.}
\item{color_intensity}{Color intensity.}
\item{hue}{Hue.}
\item{OD280_OD315_diluted}{OD280_OD315_diluted.}
\item{proline}{Proline.}
}
}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Wine}
}
\usage{
wine
}
\description{
This data is the result of a chemical analysis of wines grown in the same region in Italy but derived from three different cultivars.
The analysis determined the quantities of 13 constituents found in each of the three types of wines.
}
\keyword{datasets}
