% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_names.R
\name{preview_names}
\alias{preview_names}
\title{preview_names}
\usage{
preview_names(data, overwrite = TRUE, ...)
}
\arguments{
\item{data}{An existing dataframe that has been pulled in from the UCI
database.}

\item{overwrite}{Default is TRUE. Indicates whether new, clean column names
should replace the existing column names.}

\item{...}{Additional arguments to be passed to janitor::clean_names}
}
\description{
preview_names
}
\examples{
#show and clean names for iris
\dontrun{
iris_uci <- read_UCI("iris", "iris.data")
iris_uci_names <- preview_names(iris_uci)
}
}
