% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_UCI.R
\name{search_UCI}
\alias{search_UCI}
\title{Read files from the UCI Machine Learning Repository}
\usage{
search_UCI(
  data_type = NULL,
  data_task = NULL,
  data_area = NULL,
  min_instances = 0
)
}
\arguments{
\item{data_type}{The types of data in the dataset. Can include: Multivariate, Univariate, Text, Data-Generator, Domain-Theory, Time-Series, Spatial, Relational, Sequential, Spatio-temporal, Transactional, Image.}

\item{data_task}{The default machine learning task of the dataset. Can include: Classification, Recommender-System, Regression, Function-Learning, Relational-Learning, Clustering, Description, Causal-Discovery, Causa, Recommendation.}

\item{data_area}{The field of study of the dataset. Can include: Life Sciences, Social Sciences, Physical Sciences, CS/ Engineering, Other, Game, Business.}

\item{min_instances}{The minimum instances included the dataset. Integer can range from 1 and 63,000,000.}
}
\description{
Read files from the UCI Machine Learning Repository
}
\examples{
dataset_pref1 <- search_UCI("Time-Series", "Classification", "Life Sciences", 300)
dataset_pref2 <- search_UCI(data_type = "Text", data_area = "Business")

}
